feat: Implement boilerplate for f/setstat
This removes the Status::OpUnsupported so scp does not fail when copying over files.
It can be properly implemented at a later point in time as it really does not seem critical.

If setting file permissions is crucial to the user, it should be rather done via Windows onboard
tools rather than relying on the minimal, unix-style feature that is provided via sftp.